Split sync delta registry
