+++ service-worker.js
@@ -161,6 +161,9 @@ const APP_SHELL = [
   '/js/sync-delta-map-merge.js',
   '/js/sync-delta-scalar-merge.js',
   '/js/sync-delta-registry.js',
+  '/js/sync-delta-surfaces.js',
+  '/js/sync-delta-surface-config.js',
+  '/js/sync-delta-id.js',
   '/js/sync-delta-observability.js',
   '/js/sync-tombstones.js',
   '/js/sync-messenger.js',
